PP-7887 Fix flakiness of payment link test
The `waitForNavigation()` function returns a promise which resolves the
next time the page navigates to a new URL or reloads. As per the
puppeteer docs the promise should be obtained before the action to
change pages or reload, and then `await` for the promise to revolve to
signal that the action has completed.

Previously the test was obtaining and waiting for the promise to resolve
after it had performed the page action. This meant on ocassion the page
had already completed the navigation before `waitForNavigation()` was
invoked and the promise obtained, and the test would timeout waiting for
the navigation to complete after it had already done so.

Link to the puppeteer docs:
https://github.com/puppeteer/puppeteer/blob/v1.7.0/docs/api.md#pagewaitfornavigationoptions