PP-8387 Rationalise the workflows
The only difference between the `pull_request.yaml` and `release.yaml`
workflow configurations was that the latter had an extra step to create
the github release. In other repos such as pay-dockerfiles we use a
simple conditional when we want to perform a step only for main merges
(such as pushing the image to ECR). This reduces the amount of
configuration and prevents the files from diverging and needing to be
changed in multiple places.

- remove the `pull_request.yaml` file.
- rename the `release.yaml` to `build.yaml`
- Add a trigger to the `build.yaml` for pull requests
- Add a condition to the `build.yaml` to only create a request if its on
  the main branch.