PP-9937 Update payment link test
Update the payment link test to work with the new payment link journey.
The new journey adds in a start page and a confirm page to the journey.

Going to leave the payment link URL as it is, and just update the
existing payment links on the smoke test accounts to use the new journey
once this has deployed.