PP-12139: Fix Stripe 3DS smoke test
The smoke test was failing with `Waiting for selector "iframe.FullscreenFrame"`.
Upon inspection, the iframe we were interested is in the one with class
"ThreeDS2-challenge", not "FullscreenFrame" - see screenshot attached to this
PR. However, changing it to the correct name still didn't work. It is suspected
that this is because there is an intermediary iframe between iframe-3ds and
"ThreeDS2-challenge", so:

```
<iframe class="iframe-3ds">
  <iframe name="__privateStripeFrame0504">
    <iframe class="ThreeDS2-challenge">
    </iframe>
  </iframe>
</iframe>
```

We can't select the second iframe however because the name changes on each test
run as it's not controlled by us, and there's no id or class attributes
attached to it.

It seems the method presented in this commit works, so let's accept that and
move on.